refactor: organize common component barrel exports (#34)
