--- src/components/common/badge/index.ts
@@ -1,2 +1,0 @@
-export type { BadgeProps } from './Badge'
-export { Badge } from './Badge'
