--- src/components/common/dropdown/action-menu/index.ts
@@ -1,1 +1,0 @@
-export { ActionMenu } from './ActionMenu'
