--- src/components/common/dropdown/index.ts
@@ -1,2 +1,0 @@
-export type { DropdownOption } from './Dropdown'
-export { Dropdown } from './Dropdown'
