+++ src/components/common/layout/constants/footer.ts
@@ -1,0 +1,12 @@
+export const FOOTER_GITHUB_LINKS = [
+  {
+    href: 'https://github.com/Oz-16-Team2/oz_union_16_fe_Team2',
+    label: 'Frontend Repository',
+    text: 'Frontend',
+  },
+  {
+    href: 'https://github.com/Oz-16-Team2/oz_union_16_be_Team2',
+    label: 'Backend Repository',
+    text: 'Backend',
+  },
+]
