+++ src/components/common/layout/index.ts
@@ -1,0 +1,3 @@
+export { Footer } from './Footer'
+export { Header } from './Header'
+export { RootLayout } from './RootLayout'
