--- src/components/common/modal/base/index.ts
@@ -1,5 +1,0 @@
-export type { ModalProps } from './Modal'
-export { Modal } from './Modal'
-export { ModalContent } from './ModalContent'
-export { ModalFooter } from './ModalFooter'
-export { ModalHeader } from './ModalHeader'
