--- src/components/common/modal/confirm/index.ts
@@ -1,1 +1,0 @@
-export { ConfirmModal } from './ConfirmModal'
