--- src/components/common/modal/form/index.ts
@@ -1,1 +1,0 @@
-export { ReportFormModal } from './ReportFormModal'
