+++ src/components/common/ui/index.ts
@@ -1,0 +1,2 @@
+export type { BadgeProps } from './badge/Badge'
+export { Badge } from './badge/Badge'
