--- src/constants/footer.ts
@@ -1,14 +1,0 @@
-export const FOOTER = {
-  githubLinks: [
-    {
-      href: 'https://github.com/Oz-16-Team2/oz_union_16_fe_Team2',
-      label: 'Frontend Repository',
-      text: 'Frontend',
-    },
-    {
-      href: 'https://github.com/Oz-16-Team2/oz_union_16_be_Team2',
-      label: 'Backend Repository',
-      text: 'Backend',
-    },
-  ],
-}
